Got libvec working
