+++ Makefile
@@ -6,7 +6,7 @@
 #    By: stenner <stenner@student.42.fr>            +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2019/06/27 17:28:44 by stenner           #+#    #+#              #
-#    Updated: 2019/07/09 14:42:54 by stenner          ###   ########.fr        #
+#    Updated: 2019/07/10 13:05:05 by stenner          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -14,7 +14,7 @@ NAME = fdf
 
 LIBFT_PATH = ./libft/
 
-VEC_LIB__PATH = ./vector_lib/
+VEC_LIB_PATH = ./libvec/
 
 SRC_PATH = ./src/
 
@@ -38,8 +38,8 @@ all: $(NAME)
 
 $(NAME): $(SRCO)
 	@make -C $(LIBFT_PATH)
-	@make -C $(VEC_LIB__PATH)
-	@gcc $(FLAGS) $(SRCO) $(INC) -L $(LIBFT_PATH) -lft -o $(NAME) $(MLX_FLAGS)
+	@make -C $(VEC_LIB_PATH)
+	@gcc $(FLAGS) $(SRCO) -L $(LIBFT_PATH) -lft -L $(VEC_LIB_PATH) -lvec -o $(NAME) $(MLX_FLAGS)
 	@echo "\033[32mBinary \033[1;32m$(NAME)\033[1;0m\033[32m Created.\033[0m"
 
 $(SRC_PATH)%.o: $(SRC_PATH)%.c
@@ -51,7 +51,7 @@ clean:
 
 fclean: clean
 	@make -C $(LIBFT_PATH)/ fclean
-	@make -C $(VEC_LIB__PATH)/ fclean
+	@make -C $(VEC_LIB_PATH)/ fclean
 	@/bin/rm -rf $(NAME)
 	@echo "\033[31mBin \033[1;31m$(NAME)\033[1;0m\033[31m Removed.\033[0m"
