+++ libvec/src/matrix_rotate_x.c
@@ -1,23 +1,37 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   vector_add.c                                       :+:      :+:    :+:   */
+/*   matrix_rotate_x.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: stenner <stenner@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2019/07/09 14:36:12 by stenner           #+#    #+#             */
-/*   Updated: 2019/07/09 14:40:46 by stenner          ###   ########.fr       */
+/*   Created: 2019/07/10 11:47:23 by stenner           #+#    #+#             */
+/*   Updated: 2019/07/10 13:04:23 by stenner          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/vector_lib.h"
+#include "../includes/libvec.h"
 
-t_vector	vector_add(t_vector v1, t_vector v2)
+t_matrix	matrix_rotate_x(double theta)
 {
-	t_vector ret;
-	
-	ret.x = v1.x + v2.x;
-	ret.y = v1.y + v2.y;
-	ret.z = v1.z + v2.z;
-	return (ret);
+	t_matrix m;
+	int i;
+	int j;
+
+	i = 0;
+	j = 0;
+	(void)theta;
+	m.m[0][1] = 1;
+	// while (i < 4)
+	// {
+	// 	j = 0;
+	// 	while (j < 4)
+	// 	{
+	// 		m.m[i][j] = 0;
+	// 		j++;
+	// 	}
+	// 	i++;
+	// }
+	//m.m[0][0] = 1;
+	return m;
 }
