+++ libvec/src/vector_multiply.c
@@ -1,23 +1,23 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   vector_cross_product.c                             :+:      :+:    :+:   */
+/*   vector_multiply.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: stenner <stenner@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2019/07/10 11:15:04 by stenner           #+#    #+#             */
-/*   Updated: 2019/07/10 11:19:02 by stenner          ###   ########.fr       */
+/*   Created: 2019/07/09 14:47:07 by stenner           #+#    #+#             */
+/*   Updated: 2019/07/10 13:04:23 by stenner          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/vector_lib.h"
+#include "../includes/libvec.h"
 
-t_vector	vector_cross_product(t_vector v1, t_vector v2)
+t_vector	vector_multiply(t_vector v, double k)
 {
 	t_vector ret;
-	
-	ret.x = (v1.y * v2.z) - (v1.z * v2.y);
-	ret.y = (v1.z * v2.x) - (v1.x * v2.z);
-	ret.z = (v1.x * v2.y) - (v1.y * v2.x);
+
+	ret.x = v.x * k;
+	ret.y = v.y * k;
+	ret.z = v.z * k;
 	return (ret);
 }
